added GameController, attempted to extract some logic from GameScene / TitleScene
