+++ main/src/actions/getMoveOrAttackOrder.ts
@@ -1,11 +1,12 @@
 import { Coordinates, Direction } from '@duzh/geometry';
-import Unit from '@main/units/Unit';
-import { StayOrder } from '@main/units/orders/StayOrder';
 import { getDoor, getEnemyUnit, isBlocked } from '@main/maps/MapUtils';
-import { MoveOrder } from '@main/units/orders/MoveOrder';
+import Unit from '@main/units/Unit';
 import { AttackOrder } from '@main/units/orders/AttackOrder';
+import { MoveOrder } from '@main/units/orders/MoveOrder';
+import { StayOrder } from '@main/units/orders/StayOrder';
 import { UnitOrder } from '@main/units/orders/UnitOrder';
 
+/** @return null if blocked */
 export const getMoveOrAttackOrder = (
   unit: Unit,
   direction: Direction
